fix: Remove hard limit on number of nbd devices
Linux has a *default* limit of 16 nbd devices, but this can be changed.
Instead of hardcoding a limit or adding some sort of --max-nbd-devices
option, just use an incrementing counter and assume that we've reached the
end of  available devices when we find the first device that does not
exist.