feat: Add support for mount profiles
Add support for multiple mount profiles. A mount profile controls how
partitions are mapped to directories. The default mount profile preserves
the existing behavior of mounting each partition on a separate subdirectory
of the target directory.

The "raspberrypi" mount profile requires a device or disk image with
exactly two partitions, and mounds the second partition on the target
directory and the first partition on /boot/firmware relative to the target
directory.

The "single" mount profile requires a device or disk image with a exactly
one partition, and it mounts that partition on the target directory.